requirements/prod.txt
Pillow == 9.1.1
boto3 == 1.24.3
clearml == 1.4.1
matplotlib == 3.5.2
numpy == 1.21.2
pandas == 1.4.2
pytorch_metric_learning == 1.5.0
scikit_learn == 1.1.1
tensorboard == 2.9.0
timm == 0.4.12
torch == 1.11.0
torchvision == 0.12.0
tqdm == 4.64.0
grad-cam